+++ ios/Classes/InstabugFlutterPlugin.m
@@ -39,10 +39,12 @@ FlutterMethodChannel* channel;
           const char *type = [signature methodReturnType];
 
           if (strcmp(type, "v") != 0) {
-              void *returnVal;
+            void *returnVal;
             [inv getReturnValue:&returnVal];
             NSObject *resultSet = (__bridge NSObject *)returnVal;
             result(resultSet);
+          } else {
+            result(nil);
           }
       }
     if (!isImplemented) {