+++ lib/CrashReporting.dart
@@ -1,8 +1,11 @@
+// ignore_for_file: avoid_classes_with_only_static_members
+
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io' show Platform, exit;
-import 'package:flutter/services.dart';
+import 'dart:io' show Platform;
+
 import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
 import 'package:instabug_flutter/models/crash_data.dart';
 import 'package:instabug_flutter/models/exception_data.dart';
 import 'package:stack_trace/stack_trace.dart';
@@ -10,20 +13,18 @@ import 'package:stack_trace/stack_trace.dart';
 class CrashReporting {
   static const MethodChannel _channel = MethodChannel('instabug_flutter');
   static bool enabled = true;
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
-  }
+  static Future<String> get platformVersion async =>
+      (await _channel.invokeMethod<String>('getPlatformVersion'))!;
 
   ///Enables and disables Enables and disables automatic crash reporting.
   /// [boolean] isEnabled
-  static void setEnabled(bool isEnabled) async {
+  static Future<void> setEnabled(bool isEnabled) async {
     enabled = isEnabled;
     final List<dynamic> params = <dynamic>[isEnabled];
     await _channel.invokeMethod<Object>('setCrashReportingEnabled:', params);
   }
 
-  static void reportCrash(dynamic exception, StackTrace stack) async {
+  static Future<void> reportCrash(dynamic exception, StackTrace stack) async {
     if (kReleaseMode && enabled) {
       _reportUnhandledCrash(exception, stack);
     } else {
@@ -35,7 +36,8 @@ class CrashReporting {
   /// Reports a handled crash to you dashboard
   /// [dynamic] exception
   /// [StackTrace] stack
-  static void reportHandledCrash(dynamic exception, [StackTrace stack]) async {
+  static Future<void> reportHandledCrash(dynamic exception,
+      [StackTrace? stack]) async {
     if (stack != null) {
       _sendCrash(exception, stack, true);
     } else {
@@ -43,11 +45,12 @@ class CrashReporting {
     }
   }
 
-  static void _reportUnhandledCrash(dynamic exception, StackTrace stack) async {
+  static Future<void> _reportUnhandledCrash(
+      dynamic exception, StackTrace stack) async {
     _sendCrash(exception, stack, false);
   }
 
-  static void _sendCrash(
+  static Future<void> _sendCrash(
       dynamic exception, StackTrace stack, bool handled) async {
     final Trace trace = Trace.from(stack);
     final List<ExceptionData> frames = <ExceptionData>[];
@@ -56,7 +59,7 @@ class CrashReporting {
           trace.frames[i].uri.toString(),
           trace.frames[i].member,
           trace.frames[i].line,
-          trace.frames[i].column == null ? 0 : trace.frames[i].column));
+          trace.frames[i].column == null ? 0 : trace.frames[i].column!));
     }
     final CrashData crashData = CrashData(
         exception.toString(), Platform.operatingSystem.toString(), frames);