+++ test/network_logger_test.dart
@@ -1,27 +1,23 @@
 import 'dart:async';
 import 'dart:io';
-import 'dart:math';
 
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:instabug_flutter/NetworkLogger.dart';
 import 'package:instabug_flutter/instabug_custom_http_client.dart';
-import 'package:instabug_flutter/models/network_data.dart';
 import 'package:instabug_flutter/utils/http_client_logger.dart';
+import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-// import 'package:test_api/test_api.dart';
 
-class MockHttpClient extends Mock implements HttpClient {}
+import 'network_logger_test.mocks.dart';
 
-class MockHttpClientRequest extends Mock implements HttpClientRequest {}
-
-class MockHttpClientResponse extends Mock implements HttpClientResponse {}
-
-class MockHttpClientLogger extends Mock implements HttpClientLogger {}
-
-class MockHttpClientCredentials extends Mock implements HttpClientCredentials {}
-
+@GenerateMocks([
+  HttpClient,
+  HttpClientLogger,
+  HttpClientRequest,
+  HttpClientResponse,
+  HttpClientCredentials,
+])
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
 
@@ -29,9 +25,8 @@ void main() {
   const String url = 'https://jsonplaceholder.typicode.com';
   const int port = 8888;
   const String path = '/posts';
-  const String body = 'some request body';
 
-  InstabugCustomHttpClient instabugCustomHttpClient;
+  late InstabugCustomHttpClient instabugCustomHttpClient;
 
   setUpAll(() async {
     const MethodChannel('instabug_flutter')
@@ -52,7 +47,8 @@ void main() {
 
   test('expect instabug custom http client GET URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.getUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).getUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.getUrl(Uri.parse(url));
@@ -65,7 +61,8 @@ void main() {
 
   test('expect instabug custom http client GET to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.get(any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .get(any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.get(url, port, path);
@@ -79,7 +76,8 @@ void main() {
   test(
       'expect instabug custom http client DELETE URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.deleteUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).deleteUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.deleteUrl(Uri.parse(url));
@@ -92,7 +90,8 @@ void main() {
 
   test('expect instabug custom http client DELETE to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.delete(any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .delete(any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.delete(url, port, path);
@@ -105,7 +104,8 @@ void main() {
 
   test('expect instabug custom http client POST URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.postUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).postUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.postUrl(Uri.parse(url));
@@ -118,7 +118,8 @@ void main() {
 
   test('expect instabug custom http client POST to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.post(any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .post(any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.post(url, port, path);
@@ -131,7 +132,8 @@ void main() {
 
   test('expect instabug custom http client HEAD URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.headUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).headUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.headUrl(Uri.parse(url));
@@ -144,7 +146,8 @@ void main() {
 
   test('expect instabug custom http client HEAD to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.head(any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .head(any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.head(url, port, path);
@@ -157,7 +160,8 @@ void main() {
 
   test('expect instabug custom http client PATCH URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.patchUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).patchUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.patchUrl(Uri.parse(url));
@@ -170,7 +174,8 @@ void main() {
 
   test('expect instabug custom http client PATCH to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.patch(any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .patch(any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.patch(url, port, path);
@@ -183,7 +188,8 @@ void main() {
 
   test('expect instabug custom http client OPEN URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.openUrl(any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .openUrl(any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request =
@@ -197,7 +203,8 @@ void main() {
 
   test('expect instabug custom http client OPEN to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.open(any, any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .open(any, any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.open('GET', url, port, path);
@@ -210,7 +217,8 @@ void main() {
 
   test('expect instabug custom http client PUT URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.putUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).putUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.putUrl(Uri.parse(url));
@@ -223,7 +231,8 @@ void main() {
 
   test('expect instabug custom http client PUT to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.put(any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .put(any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.put(url, port, path);
@@ -236,7 +245,8 @@ void main() {
 
   test('expect instabug custom http client POST URL to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.postUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).postUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.postUrl(Uri.parse(url));
@@ -249,7 +259,8 @@ void main() {
 
   test('expect instabug custom http client POST to return request and log',
       () async {
-    when<dynamic>(instabugCustomHttpClient.client.post(any, any, any))
+    when<dynamic>((instabugCustomHttpClient.client as MockHttpClient)
+            .post(any, any, any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     final request = await instabugCustomHttpClient.post(url, port, path);
@@ -262,65 +273,77 @@ void main() {
 
   test('expect instabug custom http client to get client autoUncompress',
       () async {
-    when(instabugCustomHttpClient.client.autoUncompress).thenReturn(true);
+    when((instabugCustomHttpClient.client as MockHttpClient).autoUncompress)
+        .thenReturn(true);
     expect(instabugCustomHttpClient.autoUncompress, true);
   });
 
   test('expect instabug custom http client to set client autoUncompress',
       () async {
     instabugCustomHttpClient.autoUncompress = false;
-    verify(instabugCustomHttpClient.client.autoUncompress = false).called(1);
+    verify((instabugCustomHttpClient.client as MockHttpClient).autoUncompress =
+            false)
+        .called(1);
   });
 
   test('expect instabug custom http client to get client connectionTimout',
       () async {
-    when(instabugCustomHttpClient.client.connectionTimeout)
-        .thenReturn(Duration(seconds: 2));
-    expect(instabugCustomHttpClient.connectionTimeout, Duration(seconds: 2));
+    when((instabugCustomHttpClient.client as MockHttpClient).connectionTimeout)
+        .thenReturn(const Duration(seconds: 2));
+    expect(
+        instabugCustomHttpClient.connectionTimeout, const Duration(seconds: 2));
   });
 
   test('expect instabug custom http client to set client connectionTimout',
       () async {
-    instabugCustomHttpClient.connectionTimeout = Duration(seconds: 5);
-    verify(instabugCustomHttpClient.client.connectionTimeout =
-            Duration(seconds: 5))
+    instabugCustomHttpClient.connectionTimeout = const Duration(seconds: 5);
+    verify((instabugCustomHttpClient.client as MockHttpClient)
+            .connectionTimeout = const Duration(seconds: 5))
         .called(1);
   });
 
   test('expect instabug custom http client to get client idleTimeout',
       () async {
-    when(instabugCustomHttpClient.client.idleTimeout)
-        .thenReturn(Duration(seconds: 2));
-    expect(instabugCustomHttpClient.idleTimeout, Duration(seconds: 2));
+    when((instabugCustomHttpClient.client as MockHttpClient).idleTimeout)
+        .thenReturn(const Duration(seconds: 2));
+    expect(instabugCustomHttpClient.idleTimeout, const Duration(seconds: 2));
   });
 
   test('expect instabug custom http client to set client idleTimeout',
       () async {
-    instabugCustomHttpClient.idleTimeout = Duration(seconds: 5);
-    verify(instabugCustomHttpClient.client.idleTimeout = Duration(seconds: 5))
+    instabugCustomHttpClient.idleTimeout = const Duration(seconds: 5);
+    verify((instabugCustomHttpClient.client as MockHttpClient).idleTimeout =
+            const Duration(seconds: 5))
         .called(1);
   });
 
   test('expect instabug custom http client to get client maxConnectionsPerHost',
       () async {
-    when(instabugCustomHttpClient.client.maxConnectionsPerHost).thenReturn(2);
+    when((instabugCustomHttpClient.client as MockHttpClient)
+            .maxConnectionsPerHost)
+        .thenReturn(2);
     expect(instabugCustomHttpClient.maxConnectionsPerHost, 2);
   });
 
   test('expect instabug custom http client to set client maxConnectionsPerHost',
       () async {
     instabugCustomHttpClient.maxConnectionsPerHost = 5;
-    verify(instabugCustomHttpClient.client.maxConnectionsPerHost = 5).called(1);
+    verify((instabugCustomHttpClient.client as MockHttpClient)
+            .maxConnectionsPerHost = 5)
+        .called(1);
   });
 
   test('expect instabug custom http client to get client userAgent', () async {
-    when(instabugCustomHttpClient.client.userAgent).thenReturn('2');
+    when((instabugCustomHttpClient.client as MockHttpClient).userAgent)
+        .thenReturn('2');
     expect(instabugCustomHttpClient.userAgent, '2');
   });
 
   test('expect instabug custom http client to set client userAgent', () async {
     instabugCustomHttpClient.userAgent = 'something';
-    verify(instabugCustomHttpClient.client.userAgent = 'something').called(1);
+    verify((instabugCustomHttpClient.client as MockHttpClient).userAgent =
+            'something')
+        .called(1);
   });
 
   test('expect instabug custom http client to call client addClientCredentials',
@@ -350,44 +373,55 @@ void main() {
   test('expect instabug custom http client to set client authenticate',
       () async {
     final Future<bool> Function(Uri url, String scheme, String realm) f =
-        (Uri url, String scheme, String realm) {};
+        (Uri url, String scheme, String realm) async => true;
     instabugCustomHttpClient.authenticate = f;
-    verify(instabugCustomHttpClient.client.authenticate = f).called(1);
+    verify((instabugCustomHttpClient.client as MockHttpClient).authenticate = f)
+        .called(1);
   });
 
   test('expect instabug custom http client to set client authenticateProxy',
       () async {
     final Future<bool> Function(
             String host, int port, String scheme, String realm) f =
-        (String host, int port, String scheme, String realm) {};
+        (String host, int port, String scheme, String realm) async => true;
     instabugCustomHttpClient.authenticateProxy = f;
-    verify(instabugCustomHttpClient.client.authenticateProxy = f).called(1);
+    verify((instabugCustomHttpClient.client as MockHttpClient)
+            .authenticateProxy = f)
+        .called(1);
   });
 
   test(
       'expect instabug custom http client to set client badCertificateCallback',
       () async {
-    final Function(X509Certificate cert, String host, int port) f =
-        (X509Certificate cert, String host, int port) {};
+    final bool Function(X509Certificate cert, String host, int port) f =
+        (X509Certificate cert, String host, int port) => true;
     instabugCustomHttpClient.badCertificateCallback = f;
-    verify(instabugCustomHttpClient.client.badCertificateCallback = f)
+    verify((instabugCustomHttpClient.client as MockHttpClient)
+            .badCertificateCallback = f)
         .called(1);
   });
 
   test('expect instabug custom http client to call client close', () async {
     instabugCustomHttpClient.close(force: true);
-    verify(instabugCustomHttpClient.client.close(force: true)).called(1);
+    verify((instabugCustomHttpClient.client as MockHttpClient)
+            .close(force: true))
+        .called(1);
   });
 
   test('Stress test on GET URL method', () async {
-    when<dynamic>(instabugCustomHttpClient.client.getUrl(any))
+    when<dynamic>(
+            (instabugCustomHttpClient.client as MockHttpClient).getUrl(any))
         .thenAnswer((_) async => MockHttpClientRequest());
 
     for (int i = 0; i < 10000; i++) {
       await instabugCustomHttpClient.getUrl(Uri.parse(url));
     }
 
-    verify(instabugCustomHttpClient.logger.onRequest(any)).called(10000);
-    verify(instabugCustomHttpClient.logger.onResponse(any, any)).called(10000);
+    verify((instabugCustomHttpClient.logger as MockHttpClientLogger)
+            .onRequest(any))
+        .called(10000);
+    verify((instabugCustomHttpClient.logger as MockHttpClientLogger)
+            .onResponse(any, any))
+        .called(10000);
   });
 }